+++ src/synapse_migration_analyzer/__init__.py
@@ -1,3 +1,3 @@
 """Synapse Migration Analyzer."""
 
-__version__ = "3.3.0"
+__version__ = "3.4.0"